requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
fastapi==0.88.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.0.post0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0